requirements.txt
# Core
fastapi
uvicorn

# Chainlit frontend
chainlit

# DSPy framework
dspy-ai[all]

# Gemini (Google Generative AI)
google-generativeai

# Environment variable handling
python-dotenv

# HTTP for Chainlit backend calls
requests
